// quasar-project/src/types/INir.ts
export interface INir {
  code: string;
  leadCustomer: string;
  Customer: string;
  fullWorkTitle: string;
  shortWorkTitle: string;
  contractNumber: number;
  contractDate: string;
  topicNumber: number;
  workType: string;
  workPrice: number;
  startDate: string;
  endDate: string;
  stagesNumber: number;
  workPriceStructureId: number;
  responsibleExecutorLogin: string;
}